`dex-core`: moved files to `explore/` and `guard/`
